+++ src/test/kotlin/it/pullrequestfactory/io/repositories/GithubHttpWriteRepoTest.kt
@@ -56,7 +56,7 @@ class GithubHttpWriteRepoTest {
 
     private fun createGithubHttpRepo(): GithubHttpPullRequestsRepo = GithubHttpPullRequestsRepo(
             WIRE_MOCK_DEFAULT_URL + REPO_PATH,
-            "basic-auth-token",
+            "auth-token",
             QuietUI())
 
     private fun jsonFor(pr: PullRequest) =
@@ -65,7 +65,7 @@ class GithubHttpWriteRepoTest {
     private fun RequestPatternBuilder.addCommonHeaders(): RequestPatternBuilder? {
         return this
                 .withHeader("Accept", WireMock.matching("application/json"))
-                .withHeader("Authorization", WireMock.matching("Basic .*"))
+                .withHeader("Authorization", WireMock.matching("token.*"))
                 .withHeader("Content-Type", WireMock.matching("application/json"))
     }
 }